scripts/requirements.txt
# Mentark Quantum - Data Ingestion Scripts Requirements

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Hugging Face datasets
datasets>=2.14.0
huggingface-hub>=0.16.0

# Supabase
supabase>=2.0.0
python-dotenv>=1.0.0

# Pinecone (using pinecone-client for compatibility)
pinecone-client>=3.0.0,<5.0.0

# OpenAI
openai>=1.0.0

# Progress bars
tqdm>=4.65.0

# JSON handling (built-in, but listed for clarity)
# json - built-in

